chore: move service methods
