+++ internal/entrypoints/mcpserver/executor.go
@@ -26,8 +26,6 @@ func NewExecutor(
 	historyStore mcpTools.HistoryReader,
 	nodesStore mcpTools.NodesReader,
 	swarmService *swarm.Swarm,
-	serviceLogsInspector mcpTools.ServiceLogsInspector,
-	serviceSpecInspector mcpTools.ServiceSpecInspector,
 	serviceStore mcpTools.ServicesReader,
 	imageVersionResolver mcpTools.ImageVersionResolver,
 	gitRepository mcpTools.GitRepository,
@@ -44,8 +42,8 @@ func NewExecutor(
 		mcpTools.NewDockerNetworkList(swarmService.Networks),
 		mcpTools.NewDockerPluginList(swarmService.Plugins),
 		mcpTools.NewDockerSecretList(swarmService.Secrets),
-		mcpTools.NewGetServiceLogs(serviceLogsInspector),
-		mcpTools.NewGetServiceSpec(serviceSpecInspector),
+		mcpTools.NewGetServiceLogs(swarmService.Services),
+		mcpTools.NewGetServiceSpec(swarmService.Services),
 		mcpTools.NewDNSNameResolve(),
 		mcpTools.NewPingWebRoutes(serviceStore),
 		mcpTools.NewSetServiceReplicas(swarmService.Services, eventDispatcher),
