+++ internal/swarm/service_manager_test.go
@@ -1,4 +1,4 @@
-package inspector
+package swarm
 
 import (
 	"encoding/binary"
